backend/requirements.txt
fastmcp
python-dotenv
google-generativeai
pydantic
requests
fastapi
uvicorn
sqlalchemy
aiosqlite
python-multipart
openai
python-pptx